service-worker/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
greenlet==3.0.2
loguru==0.7.2
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
SQLAlchemy==2.0.23
typing_extensions==4.9.0
